ajout du -, reglage d'erreurs

+++ srcs/percent_l.c
@@ -6,7 +6,7 @@
 /*   By: kemethen <kemethen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/21 09:50:55 by kemethen          #+#    #+#             */
-/*   Updated: 2019/03/04 16:34:47 by kemethen         ###   ########.fr       */
+/*   Updated: 2019/03/28 13:28:52 by kemethen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@ size_t	percent_lx(char const *str, va_list ap, t_var *v)
 {
 	if (str[v->i + 2] == 'x')
 	{
-		hexa_low_ul(va_arg(ap, unsigned long), v);
+		hexa_low_ul(va_arg(ap, long long), v);
 		v->j = v->i + 3;
 	}
 	if (str[v->i + 2] == 'X')
 	{
-		hexa_up_ul(va_arg(ap, unsigned long), v);
+		hexa_up_ul(va_arg(ap, long long), v);
 		v->j = v->i + 3;
 	}
 	return (v->j);
@@ -40,12 +40,12 @@ size_t	percent_l(char const *str, va_list ap, t_var *v)
 		}
 		if (str[v->i + 2] == 'o')
 		{
-			percent_lo(va_arg(ap, unsigned long), v);
+			percent_lo(va_arg(ap, long long), v);
 			v->j = v->i + 3;
 		}
 		if (str[v->i + 2] == 'u')
 		{
-			percent_ld_and_li(va_arg(ap, unsigned long), v);
+			percent_ld_and_li(va_arg(ap, long long), v);
 			v->j = v->i + 3;
 		}
 		v->j = percent_lx(str, ap, v);
